refactor(syrup): split ocapn codecs into components, descriptors, passables
